Separate RH and operative close pages
