+++ src/pages/operatives/[operativeId]/work-orders/[id]/close.js
@@ -1,9 +1,9 @@
 import Meta from '@/components/Meta'
-import { getQueryProps } from '@/utils/helpers/serverSideProps'
+import CloseWorkOrder from '@/components/WorkOrders/CloseWorkOrder'
 import { OPERATIVE_ROLE } from '@/utils/user'
-import OperativeWorkOrderView from '../../../../components/WorkOrder/OperativeWorkOrderView'
+import { getQueryProps } from '@/utils/helpers/serverSideProps'
 
-const OperativeWorkOrderPage = ({ query }) => {
+const OperativeWorkOrderClosePage = ({ query }) => {
   // This page was created so users in operative and other groups
   // can access a work order via different links to take different actions.
 
@@ -11,14 +11,15 @@ const OperativeWorkOrderPage = ({ query }) => {
   // Instead the API is relied upon to check operative work order access.
   return (
     <>
-      <Meta title={`Work Order ${query.id}`} />
-      <OperativeWorkOrderView workOrderReference={query.id} />
+      <Meta title={`Close Work Order ${query.id}`} />
+
+      <CloseWorkOrder reference={query.id} />
     </>
   )
 }
 
 export const getServerSideProps = getQueryProps
 
-OperativeWorkOrderPage.permittedRoles = [OPERATIVE_ROLE]
+OperativeWorkOrderClosePage.permittedRoles = [OPERATIVE_ROLE]
 
-export default OperativeWorkOrderPage
+export default OperativeWorkOrderClosePage
